* Mac 64 bit AEffGUIEditor implementation 	* Build 32/64 bit universal binary for tutorial project 	* fix focus issues with Cocoa and text edit fields 	* fix bugs in CVUMeter

+++ trunk/vstgui/lib/platform/mac/cocoa/nsviewoptionmenu.mm
@@ -293,7 +293,7 @@ PlatformOptionMenuResult NSViewOptionMenu::popup (COptionMenu* optionMenu)
 	[cellContainer removeFromSuperviewWithoutNeedingDisplay];
 	[cellContainer release];
 	result.menu = (COptionMenu*)[nsMenu performSelector:@selector(selectedMenu)];
-	result.index = (int32_t)[nsMenu performSelector:@selector(selectedItem)];
+	result.index = (int32_t)(intptr_t)[nsMenu performSelector:@selector(selectedItem)];
 
 	return result;
 }
